requirements.txt
numpy
pandas
pillow
tensorflow>=2.10.0
opencv-python-headless
matplotlib
streamlit
streamlit-image-coordinates
pygame
# Sistem Araçları
# (os, math, json, heapq, subprocess standart kütüphanedir, eklemeye gerek yok)